LayoutTests/http/tests/site-isolation/exit-fullscreen-from-ui-process.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.waitUntilDone();
    }
</script>
<iframe src="http://localhost:8000/site-isolation/resources/exit-fullscreen-from-ui-process.html" allowfullscreen></iframe>
This test passes if it runs to completion.
