LayoutTests/http/tests/site-isolation/exit-fullscreen-from-ui-process-expected.txt

ALERT: entered fullscreen
ALERT: exited fullscreen
 This test passes if it runs to completion.
